Reworked most used words section
This used a strange algorithm to count word use - it counted every time a word was taught to someone (which is fine), but then proceeded to zero out a word's entries if it reached 50 uses?

This new system shows the top 5 most widely known words on the island and how many agents know the word.

Still to do: Add back the word parent in the most spoken words section